// src/app/services/headers-defaults.enum.ts
export enum CustomHeaderNames {
  RiotClientPlatform = "X-Riot-ClientPlatform",
  riotClientVersion = "X-RIOT-ClientVersion",
}

export enum CustomHeaderValues {
  RiotClientPlatform = "ew0KCSJwbGF0Zm9ybVR5cGUiOiAiUEMiLA0KCSJwbGF0Zm9ybU9TIjogIldpbmRvd3MiLA0KCSJwbGF0Zm9ybU9TVmVyc2lvbiI6ICIxMC4wLjE5MDQyLjEuMjU2LjY0Yml0IiwNCgkicGxhdGZvcm1DaGlwc2V0IjogIlVua25vd24iDQp9",
}